Mariano Alvira <mar@devl.org> fixes warning as error about a signed vs. unsigned comparison by casting the local unsigned variable as (long long).
git-svn-id: svn://svn.berlios.de/openocd/trunk@1600 b42882b7-edfa-0310-969c-e2dbd0fdcd60

+++ src/helper/command.c
@@ -783,7 +783,7 @@ int handle_sleep_command(struct command_context_s *cmd_ctx, char *cmd, char **ar
 	} else
 	{
 		long long then=timeval_ms();
-		while ((timeval_ms()-then)<duration)
+		while ((timeval_ms()-then)<(long long)duration)
 		{
 			target_call_timer_callbacks_now();
 			usleep(1000);
